% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_openai.R
\name{chatgpt}
\alias{chatgpt}
\title{Alias for the OpenAI Provider Function}
\usage{
chatgpt(..., .called_from = NULL)
}
\arguments{
\item{...}{Parameters passed to the appropriate OpenAI-specific function.}

\item{.called_from}{Internal routing argument; do not set manually.}
}
\description{
Alias for the OpenAI Provider Function
}
\keyword{internal}
