% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_deepseek.R
\name{deepseek}
\alias{deepseek}
\title{Deepseek Provider Function}
\usage{
deepseek(..., .called_from = NULL)
}
\arguments{
\item{...}{Parameters to be passed to the appropriate Deepseek-specific function,
such as model configuration, input text, or API-specific options.}

\item{.called_from}{An internal argument specifying which action (e.g.,
\code{chat}, \code{embed}) the function is invoked from.
This argument is automatically managed by the \code{tidyllm} verbs and should not be modified by the user.}
}
\value{
The result of the requested action, depending on the specific function invoked
(e.g., an updated \code{LLMMessage} object for \code{chat()}).
}
\description{
The \code{deepseek()} function acts as a provider interface for interacting with the Deepseek API
through \code{tidyllm}'s \code{chat()} verb.
It dynamically routes requests to deepseek-specific function. At the moment this is only
\code{deepseek_chat()}
}
