% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tokens.R
\name{estimate_tokens}
\alias{estimate_tokens}
\title{Estimate Tokens}
\usage{
estimate_tokens(.llm)
}
\arguments{
\item{.llm}{A tidyllm object}
}
\value{
The estimated number of tokens in the input list of language model responses.
}
\description{
Estimate the number of tokens for an .llm message history
}
