% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tokens.R
\name{fast_gpt4_token_estimator}
\alias{fast_gpt4_token_estimator}
\title{Fast GPT-4 Token Estimator}
\usage{
fast_gpt4_token_estimator(.text)
}
\arguments{
\item{.text}{The input text to estimate tokens from.}
}
\value{
The estimated number of tokens in the input text.
}
\description{
Estimate the number of tokens in a given text using a fast tokenization algorithm.
}
