% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{initialize_api_env}
\alias{initialize_api_env}
\title{Initialize or Retrieve API-specific Environment}
\usage{
initialize_api_env(.api_name)
}
\arguments{
\item{.api_name}{The name of the API for which to initialize or retrieve the environment}
}
\description{
This function initializes a named environment for storing rate limit information
specific to an API. It ensures that each API's rate limit data is stored separately.
}
