% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_retrieval.R
\name{last_reply_data}
\alias{last_reply_data}
\title{Get the Last Assistant Reply as Text}
\usage{
last_reply_data(.llm)
}
\arguments{
\item{.llm}{A \code{LLMMessage} object.}
}
\value{
Returns the content of the assistant's reply at the specified index, based on the following conditions:
}
\description{
A wrapper around \code{get_reply_data()} to retrieve structured data from the most recent assistant reply.
}
