% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_retrieval.R
\name{last_user_message}
\alias{last_user_message}
\title{Retrieve the Last User Message}
\usage{
last_user_message(.llm)
}
\arguments{
\item{.llm}{A \code{LLMMessage} object.}
}
\value{
The content of the last user message.
}
\description{
A wrapper around \code{get_user_message()} to retrieve the most recent user message.
}
