% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_verbs.R
\name{list_batches}
\alias{list_batches}
\title{List all Batch Requests on a Batch API}
\usage{
list_batches(.provider = getOption("tidyllm_lbatch_default"))
}
\arguments{
\item{.provider}{A function or function call specifying the language model provider and any additional parameters.
This should be a call to a provider function like \code{openai()}, \code{claude()}, etc.
You can also set a default provider function via the \code{tidyllm_lbatch_default} option.}
}
\value{
A tibble with information about the status of batch processing.
}
\description{
List all Batch Requests on a Batch API
}
