% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_verbs.R
\name{list_files}
\alias{list_files}
\title{List Files Stored on a Provider}
\usage{
list_files(.provider, ...)
}
\arguments{
\item{.provider}{A provider function call.}

\item{...}{Additional provider-specific arguments.}
}
\value{
A tibble of file metadata.
}
\description{
List Files Stored on a Provider
}
