% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_verbs.R
\name{list_models}
\alias{list_models}
\title{List Available Models for a Provider}
\usage{
list_models(.provider = getOption("tidyllm_lmodels_default"), ...)
}
\arguments{
\item{.provider}{A function or function call specifying the provider and any additional parameters.
You can also set a default provider via the \code{tidyllm_lmodels_default} option.}

\item{...}{Additional arguments to be passed to the provider-specific list_models function.}
}
\value{
A tibble containing model information.
}
\description{
The \code{list_models()} function retrieves available models from the specified provider.
}
