% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{ollama}
\alias{ollama}
\title{Call an ollama model}
\usage{
ollama(
  .llm,
  .model = "mixtral-8x7b-32768",
  .max_tokens = 1024,
  .temperature = NULL,
  .top_p = NULL,
  .frequency_penalty = NULL,
  .presence_penalty = NULL,
  .ollama_server = "http://localhost:11434",
  .timeout = 60,
  .verbose = FALSE,
  .wait = TRUE,
  .min_tokens_reset = 0L
)
}
\arguments{
\item{.llm}{An existing LLMMessage object or an initial text prompt.}

\item{.model}{The model identifier (default: "llama3").}

\item{.temperature}{Control for randomness in response generation (optional).}

\item{.top_p}{Nucleus sampling parameter (optional).}

\item{.top_k}{Top k sampling parameter (optional).}
}
\value{
Returns an updated LLMMessage object.
}
\description{
Call an ollama model
}
