% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ollama.R
\name{ollama_delete_model}
\alias{ollama_delete_model}
\title{Delete a model from the Ollama API}
\usage{
ollama_delete_model(.model, .ollama_server = "http://localhost:11434")
}
\arguments{
\item{.model}{The name of the model to delete.}

\item{.ollama_server}{The base URL of the Ollama API (default is "http://localhost:11434").}
}
\description{
This function sends a DELETE request to the Ollama API to remove a specified model.
}
