% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_list_models}
\alias{ollama_list_models}
\title{Retrieve and print model information from the Ollama API}
\usage{
ollama_list_models(.ollama_server = "http://localhost:11434")
}
\description{
This function connects to the Ollama API and retrieves and print information
about available models.
}
