% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{print_rate_limit_info}
\alias{print_rate_limit_info}
\title{Print the current rate limit information for all or a specific API}
\usage{
print_rate_limit_info(.api_name = NULL)
}
\arguments{
\item{.api_name}{(Optional) The name of the API whose rate limit info you want to print.
If not provided, the rate limit info for all APIs in the environment will be printed.}
}
\description{
This function retrieves and prints the rate limit details for the specified API,
or for all APIs stored in the .tidyllm_rate_limit_env if no API is specified.
}
