% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{rate_limit_info}
\alias{rate_limit_info}
\title{Get the current rate limit information for all or a specific API}
\usage{
rate_limit_info(.api_name = NULL)
}
\arguments{
\item{.api_name}{(Optional) The name of the API whose rate limit info you want to get
If not provided, the rate limit info for all APIs in the environment will be returned}
}
\value{
A tibble containing the rate limit information.
}
\description{
This function retrieves the rate limit details for the specified API,
or for all APIs stored in the .tidyllm_rate_limit_env if no API is specified.
}
