% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{ratelimit_from_header}
\alias{ratelimit_from_header}
\title{Extract rate limit information from API response headers}
\usage{
ratelimit_from_header(.response_headers, .api)
}
\arguments{
\item{.response_headers}{Headers from the API response}

\item{.api}{The API type ("claude", "openai","groq")}
}
\value{
A list containing rate limit information
}
\description{
Extract rate limit information from API response headers
}
