% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_verbs.R
\name{send_batch}
\alias{send_batch}
\title{Send a batch of messages to a batch API}
\usage{
send_batch(.llms, .provider = getOption("tidyllm_sbatch_default"))
}
\arguments{
\item{.llms}{A list of \code{LLMMessage} objects containing conversation histories.}

\item{.provider}{A function or function call specifying the language model provider and any additional parameters.
This should be a call to a provider function like \code{openai()}, \code{claude()}, etc.
You can also set a default provider function via the \code{tidyllm_sbatch_default} option.}
}
\value{
An updated and named list of \code{.llms} with identifiers that align with batch responses, including a \code{batch_id} attribute.
}
\description{
The \code{send_batch()} function allows you to send a list of \code{LLMMessage} objects
to an API.
It routes the input to the appropriate provider-specific batch API function.
}
