% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyllm-package.R
\docType{package}
\name{tidyllm-package}
\alias{tidyllm}
\alias{tidyllm-package}
\title{tidyllm: Tidy Integration of Large Language Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tidy interface for integrating large language model (LLM) APIs such as 'Claude', 'OpenAI', 'Gemini', 'Mistral', and local models via 'Ollama' into R workflows. The package supports text, image, audio, video, and document interactions; a unified media interface for attaching inline files or uploading to provider file stores; batch request APIs for cost-efficient large-scale processing; and a pipeline-oriented interface for seamless integration into data workflows. Web services are available at \url{https://www.anthropic.com}, \url{https://openai.com}, \url{https://aistudio.google.com/}, \url{https://mistral.ai/} and \url{https://ollama.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edubruell.github.io/tidyllm/}
  \item Report bugs at \url{https://github.com/edubruell/tidyllm/issues}
}

}
\author{
\strong{Maintainer}: Eduard Brüll \email{eduard.bruell@zew.de}

Other contributors:
\itemize{
  \item Jia Zhang [contributor]
}

}
\keyword{internal}
