% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{video_file}
\alias{video_file}
\title{Create a Video Object}
\usage{
video_file(.path)
}
\arguments{
\item{.path}{The path to the video file on disk.}
}
\value{
A \code{tidyllm_video} object.
}
\description{
Stores a reference to a local video file for use in multimodal messages.
The file is not encoded until the message is formatted for a provider.
}
