% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_voyage.R
\name{voyage}
\alias{voyage}
\title{Voyage Provider Function}
\usage{
voyage(..., .called_from = NULL)
}
\arguments{
\item{...}{Parameters to be passed to the appropriate Voyage-specific function,
such as model configuration, input text, or API-specific options.}

\item{.called_from}{An internal argument specifying which action (e.g.,
\code{embed}) the function is invoked from.
This argument is automatically managed by the \code{tidyllm} verbs and should not be modified by the user.}
}
\value{
The result of the requested action, depending on the specific function invoked
}
\description{
The \code{voyage()} function acts as a provider interface for interacting with the Voyage.ai API
through \code{tidyllm}'s verbs.
It dynamically routes requests to voyage-specific functions. At the moment this is only
\code{voyage_embed()}
}
