% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{wait_rate_limit}
\alias{wait_rate_limit}
\title{Wait for ratelimit restore times to ellapse if necessary}
\usage{
wait_rate_limit(.api_name, .min_tokens_reset)
}
\arguments{
\item{.api_name}{The name of the API for which rate limit we want to wait}

\item{.min_tokens_reset}{A token boundary at which we wish to reset (Typically should be larger than the size of the message)}
}
\description{
This function implements a standardized wait for rate limit resets
}
