% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{img}
\alias{img}
\title{Create an  Image Object}
\usage{
img(.path)
}
\arguments{
\item{.path}{The path to the image file on disk.}
}
\value{
An \code{tidyllm_image}, containing:
\itemize{
\item \code{imagepath}: The original file path
\item \code{imagename}: The basename of the image
\item \code{imagebase64}: a "data:image/...;base64,..." string
}
}
\description{
This function reads an image file from disk, encodes it in base64,
and returns a \code{tidyllm_image} object that can be used in multimodal
embedding requests.
}
