% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_retrieval.R
\name{last_reply}
\alias{last_reply}
\title{Get the Last Assistant Reply as Text}
\usage{
last_reply(.llm)
}
\arguments{
\item{.llm}{A \code{LLMMessage} object.}
}
\value{
Returns the content of the assistant's reply at the specified index, based on the following conditions:
}
\description{
A wrapper around \code{get_reply()} to retrieve the most recent assistant text reply.
}
