% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral.R
\name{mistral_embedding}
\alias{mistral_embedding}
\title{Generate Embeddings Using Mistral API}
\usage{
mistral_embedding(
  .llm,
  .model = "mistral-embed",
  .timeout = 120,
  .max_tries = 3,
  .dry_run = FALSE
)
}
\arguments{
\item{.llm}{An existing LLMMessage object (or a character vector of texts to embed)}

\item{.model}{The embedding model identifier (default: "mistral-embed").}

\item{.timeout}{Timeout for the API request in seconds (default: 120).}

\item{.max_tries}{Maximum retries to peform request}

\item{.dry_run}{If TRUE, perform a dry run and return the request object.}
}
\value{
A matrix where each column corresponds to the embedding of a message in the message history.
}
\description{
Generate Embeddings Using Mistral API
}
