% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ollama.R
\name{ollama_download_model}
\alias{ollama_download_model}
\title{Download a model from the Ollama API}
\usage{
ollama_download_model(.model, .ollama_server = "http://localhost:11434")
}
\arguments{
\item{.model}{The name of the model to download.}

\item{.ollama_server}{The base URL of the Ollama API (default is "http://localhost:11434").}
}
\description{
This function sends a request to the Ollama API to download a specified model
from Ollama's large online library of models.
}
