% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{parse_duration_to_seconds}
\alias{parse_duration_to_seconds}
\title{This internal function parses duration strings as returned by the OpenAI API}
\usage{
parse_duration_to_seconds(.duration_str)
}
\arguments{
\item{.duration_str}{A duration string.}
}
\value{
A numeric number of seconds
}
\description{
This internal function parses duration strings as returned by the OpenAI API
}
