% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_perplexity.R
\name{perplexity}
\alias{perplexity}
\title{Perplexity Provider Function}
\usage{
perplexity(..., .called_from = NULL)
}
\arguments{
\item{...}{Parameters to be passed to the appropriate Perplexity-specific function,
such as model configuration, input text, or API-specific options.}

\item{.called_from}{An internal argument specifying which action (e.g.,
\code{chat}, \code{embed}) the function is invoked from.
This argument is automatically managed by the \code{tidyllm} verbs and should not be modified by the user.}
}
\value{
The result of the requested action, depending on the specific function invoked
(e.g., an updated \code{LLMMessage} object for \code{chat()}).
}
\description{
The \code{perplexity()} function acts as a provider interface for interacting with the Perplexity API
through \code{tidyllm}'s \code{chat()} verb.
It dynamically routes requests to Perplxeity-specific function. At the moment this is only
\code{perplexity_chat()}
}
