% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyllm-package.R
\docType{package}
\name{tidyllm-package}
\alias{tidyllm}
\alias{tidyllm-package}
\title{tidyllm: Tidy Integration of Large Language Models}
\description{
A tidy interface for integrating large language model (LLM) APIs such as 'Claude', 'ChatGPT', 'Groq', and local models via 'Ollama' into R workflows. The package supports text and media-based interactions, interactive message history, stateful rate limit handling, and a tidy, pipeline-oriented interface for streamlined integration into data workflows. Web services are available at \url{https://www.anthropic.com}, \url{https://openai.com}, \url{https://groq.com}, and \url{https://ollama.com}.
}
\author{
\strong{Maintainer}: Eduard Brüll \email{eduard.bruell@zew.de}

}
\keyword{internal}
