% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyllm-package.R
\docType{package}
\name{tidyllm-package}
\alias{tidyllm}
\alias{tidyllm-package}
\title{tidyllm: Tidy Integration of Large Language Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tidy interface for integrating large language model (LLM) APIs such as 'Claude', 'Openai', 'Groq','Mistral' and local models via 'Ollama' into R workflows. The package supports text and media-based interactions, interactive message history, batch request APIs, and a tidy, pipeline-oriented interface for streamlined integration into data workflows. Web services are available at \url{https://www.anthropic.com}, \url{https://openai.com}, \url{https://groq.com}, \url{https://mistral.ai/} and \url{https://ollama.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edubruell.github.io/tidyllm/}
  \item Report bugs at \url{https://github.com/edubruell/tidyllm/issues}
}

}
\author{
\strong{Maintainer}: Eduard Brüll \email{eduard.bruell@zew.de}

Other contributors:
\itemize{
  \item Jia Zhang [contributor]
}

}
\keyword{internal}
